archive/legacy_lab/lab-main/backend/requirements.txt
asgiref==3.10.0
black==25.9.0
celery==5.4.0
click==8.3.0
coverage==7.11.0
Django==5.2.7
django-cors-headers==4.6.0
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
factory_boy==3.3.3
Faker==37.12.0
freezegun==1.5.5
gunicorn==23.0.0
iniconfig==2.3.0
isort==7.0.0
mypy==1.18.2
mypy_extensions==1.1.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.10.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-barcode==0.15.1
pytokens==0.2.0
redis==5.2.1
reportlab==4.2.5
ruff==0.14.3
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.8.2
